+++ src/main/java/racingcar/view/OutputView.java
@@ -1,6 +1,5 @@
-package racingcar.controller;
+package racingcar.view;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import racingcar.domain.Car;
